Time: 0 ms (100%), Space: 113.7 MB (61.93%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.cpp
@@ -1,12 +1,12 @@
 class Solution {
 public:
     double findMaxAverage(vector<int>& nums, int k) {
-        long long sum=0;
+        int sum=0;
         for(int i=0;i<k;i++)
         {
             sum+=nums[i];
         }
-        long long maxavg=sum;
+        int maxavg=sum;
         for(int i=k;i<nums.size();i++)
         {
             sum+=nums[i]-nums[i-k];
